Added wrapper ConvStringL18n to internationalize enum text.
