+++ src/main/resources/RBModel.properties
@@ -1,0 +1,5 @@
+TT_CASTLE.text=Castle
+TT_ROAD.text=Road
+TT_MONASTERY.text=Monastery
+TT_FIELDS.text=Fields
+TT_OTHER.text=Other
