+++ src/main/resources/RBModel_ja_JP.properties
@@ -1,0 +1,5 @@
+TT_CASTLE.text=都市
+TT_ROAD.text=道路
+TT_MONASTERY.text=修道院
+TT_FIELDS.text=草原
+TT_OTHER.text=その他
