+++ src/main/resources/RBUtil.properties
@@ -1,0 +1,4 @@
+MD_COMPASS.text=Compass
+MD_DIAMOND.text=Diamond
+MD_ROUND.text=Round
+MD_ROUNDED_SQUARE.text=Rounded Square
