+++ src/main/resources/RBUtil_ja_JP.properties
@@ -1,0 +1,4 @@
+MD_COMPASS.text=方位磁針
+MD_DIAMOND.text=ダイアモンド
+MD_ROUND.text=丸い
+MD_ROUNDED_SQUARE.text=丸い正方形
