+++ src/main/resources/RBView.properties
@@ -69,10 +69,43 @@ GSD_HEIGHT.text=Height:
 GSD_NOT_CORRECT.text= is not a valid grid sizes!
 GSD_MESSAGE.text=<html>Changes to the grid size will affect the next game. Choose values between 3 and 99.<br/>Large grids may affect the performance!</html>
 TQP_CLICK_TO_ROTATE.text=Click to rotate the tile of type 
+TQP_CASTLE_CENTER.text=Castel Center
+TQP_CASTLE_CENTER_ENTRY.text=Castel Center Entry
+TQP_CASTLE_CENTER_SIDES.text=Castel Center Side
+TQP_CASTLE_CENTER_SIDE.text=Castel Center Side
+TQP_CASTLE_EDGE.text=Castel Edge
+TQP_CASTLE_EDGE_ROAD.text=Castel Edge Road
+TQP_CASTLE_MINI.text=Castel Mini
+TQP_CASTLE_SIDES.text=Castel Sides
+TQP_CASTLE_SIDES_EDGE.text=Castel Sides Edge
+TQP_CASTLE_SIDES_EDGE_ROAD.text=Castel Side Edge Road
+TQP_CASTLE_SIDES_QUAD.text=Castel Side Quad
+TQP_CASTLE_SIDES_ROAD.text=Castel Side Road
+TQP_CASTLE_TUBE.text=Castel Tube
+TQP_CASTLE_TUBE_ENTRIES.text=Castel Tube Entryes
+TQP_CASTLE_TUBE_ENTRY.text=Castel Tube Entry
+TQP_CASTLE_WALL.text=Castel Wall
+TQP_CASTLE_WALL_CURVE_LEFT.text=Castel Wall Left
+TQP_CASTLE_WALL_CURVE_RIGHT.text=Castel wall Right
+TQP_CASTLE_WALL_ENTRY.text=Castel Wall Entry
+TQP_CASTLE_WALL_ENTRY_LEFT.text=Castel Wall Entry Left
+TQP_CASTLE_WALL_ENTRY_RIGHT.text=Castel Wall Entry Right
+TQP_CASTLE_WALL_JUNCTION.text=Castel Wall Junction
+TQP_CASTLE_WALL_ROAD.text=Castel Wall Road
+TQP_MONASTERY.text=Monastery
+TQP_MONASTERY_CASTLE.text=Monastery Castel
+TQP_MONASTERY_JUNCTION.text=Monastery Junction
+TQP_MONASTERY_ROAD.text=Monastery Road
+TQP_ROAD.text=Road
+TQP_ROAD_CROSS_SMALL.text=Roard Cross Small
+TQP_ROAD_CURVE.text=Road Carve
+TQP_ROAD_END.text=Road End
+TQP_ROAD_JUNCTION_LARGE.text=Road Junction Large
+TQP_ROAD_JUNCTION_SMALL.text=Road Junction Small
 TV_NO_TILE_TO_SELECT.text=has no Tile to select!
-TV_TOOL_TIP.text=Tile of type 
-TV_SKIP_TURN.text=Don't place tile and skip turn
 TV_ROTATE_LEFT.text=Rotate left
 TV_ROTATE_RIGHT.text=Rotate right
-ABOUT.text=\nThe board game Carcassonne is created by Klaus-Jürgen Wrede and published by Hans im Glück.\nThis computer game based on the board game is developed by Timur Sağlam.
-TITLE.text=Carcassonne
+TV_SKIP_TURN.text=Don't place tile and skip turn
+TV_TOOL_TIP.text=Tile of type 
+GML_ABOUT.text=\nThe board game Carcassonne is created by Klaus-Jürgen Wrede and published by Hans im Glück.\nThis computer game based on the board game is developed by Timur Sağlam.
+GML_TITLE.text=Carcassonne
