+++ src/main/resources/RBView_ja_JP.properties
@@ -69,10 +69,43 @@ GSD_HEIGHT.text=高さ:
 GSD_NOT_CORRECT.text= は有効なグリッド サイズではありません！
 GSD_MESSAGE.text=<html>グリッド サイズの変更は次のゲームから影響します。3 ～ 99 の値を選んでください。<br/>グリッドが大きいとゲームのパフォーマンスが低下する可能性があります！</html>
 TQP_CLICK_TO_ROTATE.text=次のタイプのタイルを回転させるためにクリックしてください: 
+TQP_CASTLE_CENTER.text=都市 中央
+TQP_CASTLE_CENTER_ENTRY.text=都市 中央 入口1箇所
+TQP_CASTLE_CENTER_SIDES.text=都市 中央 サイド
+TQP_CASTLE_CENTER_SIDE.text=都市 中央 サイド
+TQP_CASTLE_EDGE.text=都市 エッジ
+TQP_CASTLE_EDGE_ROAD.text=都市 エッジ 道路
+TQP_CASTLE_MINI.text=都市 ミニ
+TQP_CASTLE_SIDES.text=都市 サイド
+TQP_CASTLE_SIDES_EDGE.text=都市 サイド エッジ
+TQP_CASTLE_SIDES_EDGE_ROAD.text=都市 サイド エッジ 道路
+TQP_CASTLE_SIDES_QUAD.text=都市 4面 サイド
+TQP_CASTLE_SIDES_ROAD.text=都市 サイド 道路
+TQP_CASTLE_TUBE.text=都市 チューブ
+TQP_CASTLE_TUBE_ENTRIES.text=都市 チューブ 入口2箇所
+TQP_CASTLE_TUBE_ENTRY.text=都市 チューブ 入口1箇所
+TQP_CASTLE_WALL.text=都市 城壁
+TQP_CASTLE_WALL_CURVE_LEFT.text=都市 城壁 左カーブ
+TQP_CASTLE_WALL_CURVE_RIGHT.text=都市 城壁 右カーブ
+TQP_CASTLE_WALL_ENTRY.text=都市 城壁 入口
+TQP_CASTLE_WALL_ENTRY_LEFT.text=都市 城壁 入口 左向き道路
+TQP_CASTLE_WALL_ENTRY_RIGHT.text=都市 城壁 入口 右向き道路
+TQP_CASTLE_WALL_JUNCTION.text=都市 城壁	三叉路大
+TQP_CASTLE_WALL_ROAD.text=都市 城壁 道路
+TQP_MONASTERY.text=修道院
+TQP_MONASTERY_CASTLE.text=修道院 都市
+TQP_MONASTERY_JUNCTION.text=修道院 四つ辻
+TQP_MONASTERY_ROAD.text=修道院 道路
+TQP_ROAD.text=道路
+TQP_ROAD_CROSS_SMAAL.text=道路 三叉路 小
+TQP_ROAD_CURVE.text=道路 カーブ
+TQP_ROAD_END.text=道路 終端
+TQP_ROAD_JUNCTION_LARGE.text=道路 四つ辻 大
+TQP_ROAD_JUNCTION_SMALL.text=道路 三叉路 大
 TV_NO_TILE_TO_SELECT.text=選択できるタイルがありません！
 TV_TOOL_TIP.text=タイルのタイプは 
 TV_SKIP_TURN.text=タイルを置かずにターンをスキップする
 TV_ROTATE_LEFT.text=左にタイルを回す
 TV_ROTATE_RIGHT.text=右にタイルを回す
-ABOUT.text=\nボードゲームカルカソンヌは Klaus-Jürgen Wrede によって作成され、Hans im Glück により出版されました。\nボードゲームカルカソンヌをベースにしたこのコンピューター・ゲームは、Timur Sağlam によって開発されました。
-TITLE.text=カルカソンヌ
+GML_ABOUT.text=\nボードゲームカルカソンヌは Klaus-Jürgen Wrede によって作成され、Hans im Glück により出版されました。\nボードゲームカルカソンヌをベースにしたこのコンピューター・ゲームは、Timur Sağlam によって開発されました。
+GML_TITLE.text=カルカソンヌ
